ai-engine/requirements.txt
google-generativeai
pypdf